move utils to its own file, fix the export issue with common js
