+++ tests/utils/get-test-translation.js
@@ -1,0 +1,10 @@
+/**
+ * Return a test translated string.
+ *
+ * @param {string} string The string to be checked.
+ *
+ * @return {string} Test translated string.
+ */
+const getTestTranslation = ( string ) => `Translated ${ string }`;
+
+module.exports = { getTestTranslation };
